日記の追加機能の実装
