+++ ui/src/components/addDiary.tsx
@@ -1,130 +1,86 @@
 import React from "react";
 import {Mutation, MutationUpdaterFn} from "react-apollo";
 import gql from "graphql-tag";
+import { diaryFragment } from "./diary";
 
-import {RouteComponentProps} from "react-router";
-import * as H from 'history'
+import { CreateDiary } from "./__generated__/CreateDiary"
+import { GetVisitor } from "./__generated__/GetVisitor";
 
-import {diaryArticleFragment} from "./diary"
-import {PostArticle, PostArticleVariables} from "./__generated__/PostArticle"
-import { GetDiary } from "./__generated__/GetDiary";
-import {query as getDiaryQuery} from "./diary";
+import { query as getVisitorQuery } from "./index"
 
-
-
-const mutation = gql`
-    mutation PostArticle($diaryId: ID!, $title: String!, $content: String!) {
-        postArticle(diaryId: $diaryId, title: $title, content: $content) {
-            ...DiaryArticleFragment
-        }
+const createDiaryFragment = gql`
+    fragment createDiaryFragment on Diary {
+        id
+        name
     }
-    ${diaryArticleFragment}
 `;
 
+export const mutation = gql`
+    mutation CreateDiary($name: String!) {
+        createDiary(name: $name) {
+            ...createDiaryFragment
+        }
+    }
+    ${createDiaryFragment}
+`
 
-const updateArticle: (diaryId: string) => MutationUpdaterFn<PostArticle> = (diaryId) => (cache, result) => {
+export const createUpdateDiary: MutationUpdaterFn<CreateDiary> = (cache, result) => {
     const { data } = result;
-    const diary = cache.readQuery<GetDiary>({ query: getDiaryQuery, variables: {diaryId: diaryId}})       
-    if (diary && data) {
-        const name = diary.getDiary.name
-        const articles = [...diary.getDiary.articles];
-        articles.unshift(data.postArticle);
-        const newDiary = {
-            getDiary: {
-                ...diary.getDiary,
-                articles: articles,
+    const visitor = cache.readQuery<GetVisitor>({ query: getVisitorQuery })
+    if (visitor && data) {
+        const diaries = [...visitor.visitor.diaries]
+        diaries.unshift(data.createDiary)
+        const newVisitor = {
+            visitor: {
+                ...visitor.visitor,
+                diaries,
             }
-        };
-        cache.writeQuery({ query: getDiaryQuery, data: newDiary });
-    };
+        }
+        cache.writeQuery({query: getVisitorQuery, data: newVisitor})
+    }
 }
 
-interface RouteProps {
-    diaryId: string
+interface DiaryFormProps {
+    create: (name: string) => void
 }
 
-export const AddArticle: React.StatelessComponent<RouteComponentProps<RouteProps>>  = ({match, history}) => (
-    <div className="AddArticle">
-        <Mutation<PostArticle, PostArticleVariables> mutation={mutation} update={updateArticle(match.params.diaryId)}>
-            {(postArticle) => {
-                return <ArticleForm history={history}
-                post={(diaryId: string, title: string, content: string) => { postArticle({ variables: {diaryId, title, content} }) }} 
-                diaryId={match.params.diaryId}/>;
-            }}
-        </Mutation>
-    </div>
-)
-
-interface ArticleFormProps {
-    history: H.History;
-    diaryId: string;
-    post: (diaryId: string, title: string, content: string) => void;
+interface DiaryFormState {
+    name: string;
 }
-interface ArticleFormState {
-    title: string
-    content: string
-}
-
-class ArticleForm extends React.PureComponent<ArticleFormProps, ArticleFormState> {
-
+export class CreateDiaryForm extends React.PureComponent<DiaryFormProps, DiaryFormState> {
+    
     state = {
-        title: "",
-        content: "",
+        name: "",
     }
 
     private handleInput = (event: React.ChangeEvent<HTMLInputElement>) => {
         const input = event.currentTarget;
-        switch (input.name) {
-            case "title":
-                this.setState({
-                    title: input.value
-                });
-                break;
-        }
-    };
-
-    private handleTextArea = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
-        const input = event.currentTarget;
-        switch (input.name) {
-            case "title":
-                this.setState({
-                    title: input.value
-                });
-                break;
-            case "content":
-                this.setState({
-                    content: input.value
-                });
-                break;
-        }
-    };
+        this.setState({
+            name: input.value
+        })
+    }
 
     private handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-        event.preventDefault(); // これどういう意味だろう
-        this.props.post(
-            this.props.diaryId, this.state.title, this.state.content)
-        this.props.history.push(`/diaries/${this.props.diaryId}`);
+        event.preventDefault()
+        this.props.create(this.state.name)
+        this.setState({
+            name: ""
+        })
     }
 
     render() {
         return (
-            <form className="ArticleForm" onSubmit={this.handleSubmit}>
-            <div>
-                <label>Title:
-                    <input type="TEXT" name="title" value={this.state.title} onChange={this.handleInput} />
-                </label>
-            </div>
-            <div>
-                <label>Content:
-                    <textarea name="content" value={this.state.content} onChange={this.handleTextArea}> </textarea>
-                </label>
-            </div>
-            <div>
-                <button>Post</button>
-            </div>
+            <form className="CreateDiaryForm" onSubmit={this.handleSubmit}>
+                <div>
+                    <label>Diary Name:
+                        <input type="TEXT" name="name" value={this.state.name} onChange={this.handleInput} />
+                    </label>
+                </div>
+                <div>
+                    <button>Create New</button>
+                </div>
             </form>
         )
+
     }
 }
-
-
