+++ ui/src/components/index.tsx
@@ -8,6 +8,8 @@ import{ GetVisitor } from "./__generated__/GetVisitor"
 import { DeleteDiary, DeleteDiaryVariables } from "./__generated__/DeleteDiary"
 import { MutationUpdaterFn } from "apollo-client";
 
+import { createUpdateDiary, mutation as createDiary, CreateDiaryForm } from "./addDiary"
+import { CreateDiary, CreateDiaryVariables } from "./__generated__/CreateDiary"
 
 export const query = gql`
     query GetVisitor {
@@ -56,6 +58,13 @@ export const Index: React.StatelessComponent = () => (
                 }
                 const { data } = result;
                 return <>
+                    <Mutation<CreateDiary, CreateDiaryVariables> mutation={createDiary}>
+                        {(createDiary) => {
+                            return <CreateDiaryForm 
+                                    create={(name: string) =>
+                                    createDiary({ variables: {name}, update: createUpdateDiary})} />
+                        }}
+                    </Mutation>
                     <Mutation<DeleteDiary, DeleteDiaryVariables> mutation={deleteDiary}>
                         {(deleteDiary) => {
                             return <DiaryList
